Add firebase
Add firebase references.  Hook firebase email/password account creation
to signup page.

NOTE: you must create a config.js file next to the main.js file and paste
in the config object as a default export.  The file will be ignored.  This
ensures secret info is not committed to the repo.